+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/OnDemandRepairJob.java
@@ -137,6 +137,7 @@ public class OnDemandRepairJob extends ScheduledJob
     {
         return new OnDemandRepairJobView(
                 getId(),
+                myOngoingJob.getHostId(),
                 myOngoingJob.getTableReference(),
                 myRepairConfiguration,
                 getStatus(),