+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/OngoingJob.java
@@ -14,6 +14,7 @@
  */
 package com.ericsson.bss.cassandra.ecchronos.core.repair;
 
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
@@ -34,6 +35,7 @@ public class OngoingJob
     }
 
     private final UUID myJobId;
+    private final UUID myHostId;
     private final TableReference myTableReference;
     private final Map<LongTokenRange, ImmutableSet<Node>> myTokens;
     private final Set<UDTValue> myRepairedTokens;
@@ -47,6 +49,7 @@ public class OngoingJob
     {
         myOnDemandStatus = builder.onDemandStatus;
         myJobId = builder.jobId == null ? UUID.randomUUID() : builder.jobId;
+        myHostId = builder.hostId;
         myTableReference = builder.tableReference;
         myReplicationState = builder.replicationState;
         myTokens = myReplicationState.getTokenRangeToReplicas(myTableReference);
@@ -64,6 +67,11 @@ public class OngoingJob
     public UUID getJobId()
     {
         return myJobId;
+    }
+
+    public UUID getHostId()
+    {
+        return myHostId;
     }
 
     public Status getStatus()
@@ -101,8 +109,43 @@ public class OngoingJob
 
     public boolean hasTopologyChanged()
     {
-    	return !myTokens.equals(myReplicationState.getTokenRangeToReplicas(myTableReference))
+        return !myTokens.equals(myReplicationState.getTokenRangeToReplicas(myTableReference))
                 || (myTokenHash != null && (myTokenHash != myTokens.keySet().hashCode() && myTokenHash != myTokens.hashCode()));
+    }
+
+    public void startClusterWideJob()
+    {
+        Map<LongTokenRange, ImmutableSet<Node>> allTokenRanges = myReplicationState.getTokenRanges(myTableReference);
+        Map<Node, Set<LongTokenRange>> repairedRangesPerNode = new HashMap<>();
+        Map<Node, Set<LongTokenRange>> remainingRangesPerNode = new HashMap<>();
+        for (Map.Entry<LongTokenRange, ImmutableSet<Node>> range : allTokenRanges.entrySet())
+        {
+            LongTokenRange rangeForNodes = range.getKey();
+            Set<Node> nodes = range.getValue();
+            boolean rangeRepaired = myTokens.containsKey(rangeForNodes);
+            for (Node node: nodes)
+            {
+                if (rangeRepaired)
+                {
+                    repairedRangesPerNode.computeIfAbsent(node, (k) -> new HashSet<>()).add(rangeForNodes);
+                }
+                else
+                {
+                    remainingRangesPerNode.computeIfAbsent(node, (k) -> new HashSet<>()).add(rangeForNodes);
+                    rangeRepaired = true; // We only want one node to repair the range
+                }
+            }
+        }
+        for (Map.Entry<Node, Set<LongTokenRange>> replicaWithRanges: remainingRangesPerNode.entrySet())
+        {
+            Node node = replicaWithRanges.getKey();
+            Set<LongTokenRange> remainingRanges = replicaWithRanges.getValue();
+            Set<LongTokenRange> repairedRanges = repairedRangesPerNode.get(node);
+            Set<LongTokenRange> allTokensForNode = new HashSet<>();
+            allTokensForNode.addAll(remainingRanges);
+            allTokensForNode.addAll(repairedRanges);
+            myOnDemandStatus.addNewJob(node.getId(), myJobId, myTableReference, allTokensForNode.hashCode(), repairedRanges);
+        }
     }
 
     public void finishJob()
@@ -118,6 +161,7 @@ public class OngoingJob
     public static class Builder
     {
         private UUID jobId = null;
+        private UUID hostId;
         private TableReference tableReference;
         private Set<UDTValue> repairedTokens = new HashSet<>();
         private OnDemandStatus onDemandStatus;
@@ -154,6 +198,12 @@ public class OngoingJob
         public Builder withReplicationState(ReplicationState replicationState)
         {
             this.replicationState = replicationState;
+            return this;
+        }
+
+        public Builder withHostId(UUID hostId)
+        {
+            this.hostId = hostId;
             return this;
         }