+++ ecchronos-binary/src/bin/ecctool.py
@@ -84,6 +84,8 @@ def add_repairs_subcommand(sub_parsers):
     parser_repairs.add_argument("-l", "--limit", type=int,
                                 help="Limit the number of tables or virtual nodes printed (-1 to disable)",
                                 default=-1)
+    parser_repairs.add_argument("--hostid", type=str,
+                                help='Show repairs for the specified host id')
 
 def add_schedules_subcommand(sub_parsers):
     parser_schedules = sub_parsers.add_parser("schedules",
@@ -135,6 +137,8 @@ def add_run_repair_subcommand(sub_parsers):
     parser_trigger_repair.add_argument("-u", "--url", type=str,
                                        help="The host to connect to with the format (http://<host>:port)",
                                        default=None)
+    parser_trigger_repair.add_argument("--local", action='store_true',
+                                       help='repair will run for the local node, disabled by default', default=False)
     required_args = parser_trigger_repair.add_argument_group("required arguments")
     required_args.add_argument("-k", "--keyspace", type=str,
                                help="Keyspace where the repair should be triggered", required=True)
@@ -228,7 +232,7 @@ def repairs(arguments):
     request = rest.V2RepairSchedulerRequest(base_url=arguments.url)
     printer = table_printer_v2
     if arguments.id:
-        result = request.get_repair(job_id=arguments.id)
+        result = request.get_repair(job_id=arguments.id, host_id=arguments.hostid)
         if result.is_successful():
             printer.print_repairs(result.data, arguments.limit)
         else:
@@ -237,13 +241,13 @@ def repairs(arguments):
         if not arguments.keyspace:
             print("Must specify keyspace")
             sys.exit(1)
-        result = request.list_repairs(keyspace=arguments.keyspace, table=arguments.table)
+        result = request.list_repairs(keyspace=arguments.keyspace, table=arguments.table, host_id=arguments.hostid)
         if result.is_successful():
             printer.print_repairs(result.data, arguments.limit)
         else:
             print(result.format_exception())
     else:
-        result = request.list_repairs(keyspace=arguments.keyspace)
+        result = request.list_repairs(keyspace=arguments.keyspace, host_id=arguments.hostid)
         if result.is_successful():
             printer.print_repairs(result.data, arguments.limit)
         else:
@@ -269,9 +273,9 @@ def repair_config(arguments):
 def run_repair(arguments):
     request = rest.V2RepairSchedulerRequest(base_url=arguments.url)
     printer = table_printer_v2
-    result = request.post(keyspace=arguments.keyspace, table=arguments.table)
+    result = request.post(keyspace=arguments.keyspace, table=arguments.table, local=arguments.local)
     if result.is_successful():
-        printer.print_repair(result.data)
+        printer.print_repairs(result.data)
     else:
         print(result.format_exception())