+++ ecchronos-binary/src/pylib/ecchronoslib/rest.py
@@ -162,9 +162,10 @@ class V2RepairSchedulerRequest(RestRequest):
 
         return result
 
-    def get_repair(self, job_id):
+    def get_repair(self, job_id, host_id=None):
         request_url = V2RepairSchedulerRequest.v2_repair_id_status_url.format(job_id)
-
+        if host_id:
+            request_url += "?hostId={0}".format(host_id)
         result = self.request(request_url)
         if result.is_successful():
             result = result.transform_with_data(new_data=[Repair(x) for x in result.data])
@@ -186,13 +187,16 @@ class V2RepairSchedulerRequest(RestRequest):
 
         return result
 
-    def list_repairs(self, keyspace=None, table=None):
+    def list_repairs(self, keyspace=None, table=None, host_id=None):
         request_url = V2RepairSchedulerRequest.v2_repair_status_url
-
-        if keyspace and table:
-            request_url = "{0}?keyspace={1}&table={2}".format(request_url, keyspace, table)
-        elif keyspace:
+        if keyspace:
             request_url = "{0}?keyspace={1}".format(request_url, keyspace)
+            if table:
+                request_url += "&table={0}".format(table)
+            if host_id:
+                request_url += "&hostId={0}".format(host_id)
+        elif host_id:
+            request_url += "?hostId={0}".format(host_id)
 
         result = self.request(request_url)
 
@@ -201,11 +205,13 @@ class V2RepairSchedulerRequest(RestRequest):
 
         return result
 
-    def post(self, keyspace=None, table=None):
+    def post(self, keyspace=None, table=None, local=False):
         request_url = V2RepairSchedulerRequest.v2_repair_trigger_url.format(keyspace, table)
+        if local:
+            request_url += "&isLocal=true"
         result = self.request(request_url, 'POST')
         if result.is_successful():
-            result = result.transform_with_data(Repair(result.data))
+            result = result.transform_with_data(new_data=[Repair(x) for x in result.data])
         return result