+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Version 3.0.0
 
+* Add support for clusterwide repairs - Issue #299
 * Add custom HTML error page
 * Make fault reporter pluggable
 * Fix JMX URI validation with new Java versions - Issue #306